tier3shopping mostly done
problem now is that buying  components of a tier3item is  much more complicated for a few reasons but one of them being  that I don't allow the shopper to buy more than one of any single tier3 but recipes for tier3  might have f.x. two ruby crystals in it. But i also need to think about what if he doesn't have enough gold for most EffHP item in recipe what  he  should try to buy then and if whatever he has left  to buy of the recipe might have its own tier2  recipe that i need to consider